Added Phase 4 dashboard implementation
